+++ source/register-content-script-shim.ts
